+++ packages/config/tests/normalize/tests.js
@@ -14,22 +14,18 @@ test('Some properties can be capitalized', async (t) => {
 
 test('Some properties can be capitalized even when merged with defaultConfig', async (t) => {
   const defaultConfig = JSON.stringify({
-    build: {
-      base: 'baseDefault',
-      command: 'gulp build default',
-      functions: 'functionsDefault',
-      edge_handlers: 'edgeHandlersDefault',
-      ignore: 'doIgnoreDefault',
-      publish: 'publishDefault',
-      environment: { TEST: 'testDefault' },
-      processing: { css: { bundle: false } },
+    Build: {
+      Base: 'base',
+      Command: 'gulp build default',
+      Functions: 'functions',
+      Edge_handlers: 'edgeHandlers',
+      Ignore: 'doIgnore',
+      Publish: 'publish',
+      Environment: { TEST: 'test' },
+      Processing: { css: { bundle: false } },
     },
   })
-  await runFixture(t, 'props_case', { flags: { defaultConfig } })
-})
-
-test('Some properties can be capitalized even when merged with contexts', async (t) => {
-  await runFixture(t, 'props_case_context', { flags: { context: 'testContext', branch: 'testBranch' } })
+  await runFixture(t, 'props_case_default_config', { flags: { defaultConfig } })
 })
 
 test('Does not add build.commandOrigin config if there are none', async (t) => {
@@ -91,4 +87,29 @@ test('Normalizes function configurations defined at different levels', async (t)
 
 test('Handles function configuration objects for functions with the same name as one of the configuration properties', async (t) => {
   await runFixture(t, 'function_config_ambiguous')
+})
+
+test('Merges plugins in netlify.toml and defaultConfig', async (t) => {
+  const defaultConfig = JSON.stringify({
+    plugins: [
+      {
+        package: 'netlify-plugin-test',
+        inputs: {
+          boolean: true,
+          unset: true,
+          array: ['a', 'b'],
+          object: { prop: true },
+        },
+      },
+    ],
+  })
+  await runFixture(t, 'merge_netlify_toml_default', { flags: { defaultConfig } })
+})
+
+test('Merges context-specific plugins', async (t) => {
+  await runFixture(t, 'merge_netlify_toml_context')
+})
+
+test('Context-specific plugins config is last in merged array', async (t) => {
+  await runFixture(t, 'merge_netlify_toml_context_last')
 })