+++ packages/functions-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+### [1.3.23](https://www.github.com/netlify/build/compare/functions-utils-v1.3.22...functions-utils-v1.3.23) (2021-04-07)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @netlify/zip-it-and-ship-it to ^3.2.0 ([#2552](https://www.github.com/netlify/build/issues/2552)) ([26b379e](https://www.github.com/netlify/build/commit/26b379e10feb5ee26c3fd426df05a21c0eafb4f1))
+
+### [1.3.22](https://www.github.com/netlify/build/compare/functions-utils-v1.3.21...functions-utils-v1.3.22) (2021-04-06)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @netlify/zip-it-and-ship-it to ^3.1.0 ([#2536](https://www.github.com/netlify/build/issues/2536)) ([fc694be](https://www.github.com/netlify/build/commit/fc694be35a186ef362c3e1e541d0e4c65af7109d))
+
 ### [1.3.21](https://www.github.com/netlify/build/compare/v1.3.20...v1.3.21) (2021-03-18)